ApiTickectsPostMethod

+++ ProjFinalCinelAirClient/Data/SeedDb.cs
@@ -163,7 +163,7 @@ namespace ProjFinalCinelAirClient.Data
             });
         }
 
-        private void Add_Mile_Status(decimal miles, DateTime validity, string userId)
+        private void Add_Mile_Status(int miles, DateTime validity, string userId)
         {
             _context.Mile_Status.Add(new Mile_Status
             {
@@ -174,7 +174,7 @@ namespace ProjFinalCinelAirClient.Data
         }
 
 
-        private void Add_Mile_Bonus(decimal miles, DateTime validity, string userId)
+        private void Add_Mile_Bonus(int miles, DateTime validity, string userId)
         {
             _context.Mile_Bonus.Add(new Mile_Bonus
             {
@@ -185,7 +185,7 @@ namespace ProjFinalCinelAirClient.Data
         }
 
 
-        private void Add_Rate(string description, decimal percentage)
+        private void Add_Rate(string description, int percentage)
         {
             _context.Rate.Add(new Rate
             {
